edit

+++ printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/28 14:52:50 by hkubo             #+#    #+#             */
-/*   Updated: 2021/03/23 11:57:30 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/03/24 09:57:54 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ int		str_to_num_field(const char *arg, int *i, va_list *ap, t_plist *f_list);
 int		str_to_num_pre(const char *arg, int *i, va_list *ap, int k);
 t_plist	init_plist(void);
 int		get_len_long(unsigned long n);
-int		field_no_precision_sub(int num, char *str, t_plist flag_list, int i);
+int		field_no_precision_sub(long num, char *str, t_plist flag_list, int i);
 char	*field_precision_sub(int num, char *str, t_plist flag_list, int i);
 int		no_field_int_sub(t_plist flag_list, int num, int digit);
 int		print_int_len(t_plist flag_list, char *str_num, int keta, int num);
